requirements.txt
streamlit
Google-Images-Search
openai
pandas
requests
Pillow
bs4
streamlit-extras
numpy
torch
plotly-express
protobuf==3.20.0